Exchange action done

+++ Revolut-test/Exchange/Submodules/Currency/Entity/CoreData/Currency+CoreDataProperties.h
@@ -1,3 +1,11 @@
+//
+//  Currency+CoreDataProperties.h
+//  Revolut-test
+//
+//  Created by Alexey Getman on 22/02/2017.
+//  Copyright © 2017 AGG. All rights reserved.
+//  This file was automatically generated and should not be edited.
+//
 
 #import "Currency+CoreDataClass.h"
 
@@ -8,7 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSFetchRequest<Currency *> *)fetchRequest;
 
-@property (nonatomic) int32_t amount;
+@property (nonatomic) double amount;
 @property (nullable, nonatomic, copy) NSString *symbol;
 @property (nullable, nonatomic, copy) NSString *title;
