+++ Revolut-test/Exchange/Submodules/Currency/Entity/PONSO/PONSO_Currency.h
@@ -3,7 +3,7 @@
 
 @interface PONSO_Currency : NSObject
 
-@property (assign, nonatomic) int32_t   amount;
+@property (assign, nonatomic) double   amount;
 @property (strong, nonatomic) NSString  *symbol;
 @property (strong, nonatomic) NSString  *title;
